+++ projects/components/src/lib/form-field/index.ts
@@ -1,5 +1,5 @@
+export * from '../form';
 export * from './form-field';
-export * from './au-form';
 export {
   AuFieldIdGenerator,
   AU_FIELD_AUTO_ID_PATTERN,